+++ examples/app-router/app/layout.tsx
@@ -6,9 +6,11 @@ import Link from "next/link";
 
 const inter = Inter({ subsets: ["latin"] });
 const forcedPages: ForcedPage[] = [
-  [/forced-color-scheme\/dark/, { colorScheme: "dark" }],
-  [/forced-color-scheme\/light/, { colorScheme: "light" }],
-  ...[...darkThemes, ...lightThemes].map(th => [new RegExp(`themed-page/${th}`), { theme: "light" }] as ForcedPage),
+  { pathMatcher: /forced-color-scheme\/dark/, props: { forcedColorScheme: "dark" } },
+  { pathMatcher: /forced-color-scheme\/light/, props: { forcedColorScheme: "light" } },
+  ...[...darkThemes, ...lightThemes].map(
+    theme => ({ pathMatcher: new RegExp(`themed-page/${theme}`), props: { forcedTheme: theme } }) as ForcedPage,
+  ),
 ];
 
 export default function RootLayout({ children }: { children: React.ReactNode }): JSX.Element {