+++ lib/nextjs-themes/src/client/theme-switcher/theme-switcher.tsx
@@ -2,12 +2,13 @@ import * as React from "react";
 import { useEffect } from "react";
 import type { ColorSchemeType } from "../../store";
 import { useTheme } from "../../store";
-import { resolveThemeFromColorScheme } from "../../utils";
+import { resolveTheme } from "../../utils";
 
 export interface ThemeSwitcherProps {
   forcedTheme?: string;
   forcedColorScheme?: ColorSchemeType;
   targetSelector?: string;
+  themeTransition?: string;
 }
 
 export function ThemeSwitcher(props: ThemeSwitcherProps) {
@@ -27,30 +28,18 @@ export function useThemeSwitcher(props: ThemeSwitcherProps) {
 
   useEffect(() => {
     const themeState = useTheme.getState();
-    const { colorSchemePref, forcedTheme, forcedColorScheme, setResolved } = themeState;
-    const resolvedForcedTheme = props.forcedTheme === undefined ? forcedTheme : props.forcedTheme;
-    const resolvedForcedColorScheme =
-      props.forcedColorScheme === undefined ? forcedColorScheme : props.forcedColorScheme;
-    const colorScheme = resolvedForcedColorScheme === undefined ? colorSchemePref : resolvedForcedColorScheme;
-
     const media = matchMedia("(prefers-color-scheme: dark)");
     const updateTheme = () => {
-      const restoreTransitions = disableAnimation();
+      const restoreTransitions = disableAnimation(props.themeTransition);
 
-      const newTheme =
-        resolvedForcedTheme !== undefined
-          ? resolvedForcedTheme
-          : resolveThemeFromColorScheme({ ...themeState, colorSchemePref: colorScheme }, media.matches);
-
-      setResolved({ resolvedTheme: newTheme, resolvedColorScheme: colorScheme });
-      updateDOM({ newTheme, colorScheme, media }, props.targetSelector);
+      const resolvedData = resolveTheme(media.matches, themeState, props);
+      themeState.setResolved(resolvedData);
+      updateDOM(resolvedData, media.matches, props.targetSelector);
 
       restoreTransitions();
     };
 
-    if (colorScheme === "system") media.addEventListener("change", updateTheme);
-    else media.removeEventListener("change", updateTheme);
-
+    media.addEventListener("change", updateTheme);
     updateTheme();
     return () => {
       media.removeEventListener("change", updateTheme);
@@ -58,30 +47,46 @@ export function useThemeSwitcher(props: ThemeSwitcherProps) {
   }, [props.forcedColorScheme, props.forcedTheme, props.targetSelector, ...depArray]);
 }
 
-interface UpdateDOMProps {
-  newTheme: string;
-  media: MediaQueryList;
-  colorScheme: ColorSchemeType;
+export interface DataProps {
+  className: string;
+  "data-th"?: string;
+  "data-theme"?: string;
+  "data-color-scheme"?: "dark" | "light";
+  "data-csp"?: ColorSchemeType /** color-scheme-preference */;
 }
 
-function updateDOM({ newTheme, colorScheme, media }: UpdateDOMProps, targetSelector?: string) {
+export interface UpdateProps {
+  resolvedTheme: string;
+  resolvedColorScheme: "dark" | "light";
+  resolvedColorSchemePref: ColorSchemeType;
+  th: string;
+}
+
+function updateDOM(
+  { resolvedTheme, resolvedColorScheme, resolvedColorSchemePref, th }: UpdateProps,
+  isSystemDark: boolean,
+  targetSelector?: string,
+) {
   [document.querySelector(targetSelector || "#nextjs-themes"), document.documentElement].forEach(target => {
-    if (target) target.className = `${newTheme} cs-${colorScheme}`; /** cs == color-scheme */
-    target?.setAttribute("data-theme", newTheme);
-    target?.setAttribute("data-color-scheme", colorScheme);
+    /** ensuring that class 'dark' is always present when dark color scheme is applied to support Tailwind  */
+    if (target)
+      target.className = `${resolvedColorScheme} theme-${resolvedTheme} th-${th} csp-${resolvedColorSchemePref}`;
+    target?.setAttribute("data-th", th);
+    target?.setAttribute("data-theme", resolvedTheme);
+    target?.setAttribute("data-color-scheme", resolvedColorScheme);
+    target?.setAttribute("data-csp", resolvedColorSchemePref); /** color-scheme-preference */
   });
 
   /** store system preference for computing data-theme on server side */
-  document.cookie = `data-color-scheme-system=${media.matches ? "dark" : "light"}`;
+  document.cookie = `data-color-scheme-system=${isSystemDark ? "dark" : "light"}`;
 }
 
 // todo: customizable transition time
-const disableAnimation = () => {
+const disableAnimation = (themeTransition = "none") => {
   const css = document.createElement("style");
+  const transition = `transition: ${themeTransition} !important;`;
   css.appendChild(
-    document.createTextNode(
-      `*{-webkit-transition:none!important;-moz-transition:none!important;-o-transition:none!important;-ms-transition:none!important;transition:none!important}`,
-    ),
+    document.createTextNode(`*{-webkit-${transition}-moz-${transition}-o-${transition}-ms-${transition}${transition}}`),
   );
   document.head.appendChild(css);